Refactor Viewer access patterns.
Viewer is accessed in the Nav component via composed GraphQL Fragments. For the rest of the app access is provided in the GlobalQuery context, and a useViewer hook accesses the fragment provided in the context.
Fix bug when viewer is null from graphql response.
Show login in ViewerRsvpStatus if not logged in.